Move the search bar to the data-explorte
